Added View menu for organization
New view menu contains particle count, background selector, and zoom modifier buttons.

+++ Parme.Editor/InputHandler.cs
@@ -103,7 +103,7 @@ namespace Parme.Editor
                 var scrollChange = _currentMouseState.ScrollWheelValue - _previousMouseState.ScrollWheelValue;
                 if (scrollChange != 0)
                 {
-                    const float scrollZoomModifier = 0.1f;
+                    const decimal scrollZoomModifier = 0.1m;
                     _applicationState.Zoom = scrollChange > 0
                         ? _applicationState.Zoom + scrollZoomModifier
                         : _applicationState.Zoom - scrollZoomModifier;
